+++ database/factories/DomainFactory.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Uccello\Core\Database\Factories;
+
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Uccello\Core\Models\Domain;
+
+class DomainFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Domain::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->company(),
+        ];
+    }
+}