+++ database/migrations/2021_04_25_173125_create_profile_module.php
@@ -5,7 +5,7 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Uccello\Core\Models\Module;
 
-class CreateDomainModule extends Migration
+class CreateProfileModule extends Migration
 {
     /**
      * Run the migrations.
@@ -20,33 +20,28 @@ class CreateDomainModule extends Migration
 
     private function createTable()
     {
-        Schema::create(config('uccello.database.table_prefix').'domains', function (Blueprint $table) {
+        $prefix = config('uccello.database.table_prefix');
+        Schema::create($prefix.'profiles', function (Blueprint $table) use ($prefix) {
             $table->id();
             $table->string('name');
-            $table->string('slug')->unique();
-            $table->foreignId('parent_id')->nullable()->constrained($table->getTable());
-            $table->string('path', 255)->nullable();
-            $table->integer('level')->default(0);
+            $table->foreignId('domain_id')->constrained($prefix.'domains');
             $table->json('data')->nullable();
             $table->timestamps();
             $table->softDeletes();
-
-            // Index
-            $table->index(['path', 'parent_id', 'level']);
         });
     }
 
     private function createModule()
     {
         Module::create([
-            'name' => 'domain',
+            'name' => 'profile',
             'data' => [
                 'package' => 'uccello/uccello',
-                'model' => \Uccello\Core\Models\Domain::class,
+                'model' => \Uccello\Core\Models\Profile::class,
                 'admin' => true,
                 'required' => true,
                 'structure' => [
-                    'icon' => 'domain',
+                    'icon' => 'lock',
                     'tabs' => [
                         [
                             'name' => 'main',
@@ -60,17 +55,10 @@ class CreateDomainModule extends Migration
                                             'uitype' => 'string',
                                             'displaytype' => 'everywhere',
                                             'required' => true,
-                                            'rules' => [
-                                                'regex:/(?!^\d+$)^.+$/',
-                                                'unique:'.config('uccello.database.table_prefix').'domains,name,%id%'
-                                            ]
                                         ],
                                         [
-                                            'name' => 'parent',
-                                            'uitype' => [
-                                                'name' => 'entity',
-                                                'module' => 'domain',
-                                            ],
+                                            'name' => 'description',
+                                            'uitype' => 'string',
                                             'displaytype' => 'everywhere',
                                         ],
                                     ]
@@ -99,7 +87,7 @@ class CreateDomainModule extends Migration
                         [
                             'name' => 'all',
                             'type' => 'list',
-                            'columns' => ['name', 'parent'],
+                            'columns' => ['name', 'description'],
                             'default' => true,
                             'readonly' => true,
                         ],
@@ -116,8 +104,8 @@ class CreateDomainModule extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists(config('uccello.database.table_prefix').'domains');
+        Schema::dropIfExists(config('uccello.database.table_prefix').'profiles');
 
-        Module::where('name', 'domain')->delete();
+        Module::where('name', 'profile')->delete();
     }
 }