+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.0] - 2021-04-25
 ### Added
-- Uccello now uses repository `Uccello\Core\Models\Repositories\ModuleRepository` to retrieve module records.
+- Uccello now uses repository `Uccello\Core\Repositories\RecordRepository` to retrieve module records.
 
 ### Changed
 - Controllers uses trait `Uccello\Core\Support\Traits\UccelloController` instead of extends an overrided controller class.